+++ Include/Intel/IndustryStandard/ProcessorInfo.h
@@ -250,6 +250,9 @@ typedef enum {
 #define CPU_MODEL_ALDERLAKE_S     0x97
 #define CPU_MODEL_RAPTORLAKE_S    0xB7 ///< Raptor Lake B0 stepping
 #define CPU_MODEL_RAPTORLAKE_HX   0xBF ///< Raptor Lake C0 stepping
+#define CPU_MODEL_ARROWLAKE_S     0xC6 ///< desktop ArrowLake
+#define CPU_MODEL_ARROWLAKE_HX    0xC5
+#define CPU_MODEL_ARROWLAKE_U     0xB5
 
 #define AMD_CPU_FAMILY          0xF
 #define AMD_CPU_EXT_FAMILY_0FH  0x0