+++ Changelog.md
@@ -3,6 +3,8 @@ OpenCore Changelog
 #### v1.0.3
 - Fixed support for `AMD_CPU_EXT_FAMILY_1AH`, thx @Shaneee
 - Fixed EHCI handoff logic in OpenDuet, causing older machines to hang at start
+- Added Arrow Lake CPU detection
+- Fixed Raptor Lake CPU detection
 
 #### v1.0.2
 - Fixed error in macrecovery when running headless, thx @mkorje